修正流量统计不支持统计超过2TB数据的的问题（#369）

+++ TrafficMonitor/CalendarHelper.h
@@ -5,12 +5,12 @@
 struct DayTraffic
 {
 	int day;
-	int up_traffic;
-	int down_traffic;
+	__int64 up_traffic;
+	__int64 down_traffic;
 	bool mixed;
 	CRect rect;
 
-	int traffic() const
+	__int64 traffic() const
 	{
 		return up_traffic + down_traffic;
 	}
